# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: clipp'n'k API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/roles/sync:
    post:
      operationId: syncRoles
      tags: [auth]
      summary: Re-sync the current user's Discord roles on demand
      responses:
        "200":
          description: Updated current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Me:
      type: object
      required:
        [
          id,
          discordId,
          username,
          avatarUrl,
          isAdmin,
          usedStorageBytes,
          quotaStorageBytes,
          maxUploadBytes,
          roles,
          banned,
          githubUsername,
          githubStarBonusGranted,
        ]
      properties:
        id: { type: integer }
        discordId: { type: string }
        username: { type: string }
        avatarUrl: { type: ["string", "null"] }
        isAdmin: { type: boolean }
        usedStorageBytes: { type: integer }
        quotaStorageBytes: { type: integer }
        maxUploadBytes:
          type: integer
          description: >-
            Effective per-file upload size limit for this user (role-aware),
            used by the upload dialog to filter oversized files client-side.
        roles:
          type: array
          items: { type: string }
          description: Discord role IDs this user currently holds in the configured guild.
        banned:
          type: boolean
          description: Whether an admin has banned this account.
        githubUsername:
          type: ["string", "null"]
          description: GitHub login of the linked account, null if not connected.
        githubStarBonusGranted:
          type: boolean
          description: Whether the +1 GB star bonus has already been permanently granted.
...